PowerShell script to build V8 11.6.189.22.

+++ BuildV8FromSource.ps1
@@ -1,5 +1,5 @@
-﻿$V8Version="11.5.150.16"
-# Previous 10.4.132.20
+﻿$V8Version="11.6.189.22"
+# Previous 11.5.150.16
 Function GetV8DebugArgs
 {
 @'
@@ -8,6 +8,7 @@ Function GetV8DebugArgs
 symbol_level=2
 icu_use_data_file=false
 use_custom_libcxx=false
+is_clang=true
 is_component_build=true
 is_debug=true
 is_official_build=false
@@ -21,12 +22,12 @@ v8_enable_debugging_features=true
 v8_enable_disassembler=true
 v8_enable_object_print=true
 v8_enable_pointer_compression=false
+v8_enable_webassembly=false
 v8_generate_external_defines_header=true
 v8_optimized_debug=false
 v8_postmortem_support=true
 v8_imminent_deprecation_warnings=false
 v8_deprecation_warnings=false
-use_cxx17=true
 cppgc_enable_young_generation=true
 '@
 }
@@ -38,6 +39,7 @@ Function GetV8ReleaseArgs
 symbol_level=2
 icu_use_data_file=false
 use_custom_libcxx=false
+is_clang=true
 is_component_build=true
 is_debug=false
 is_official_build=false
@@ -51,12 +53,12 @@ v8_enable_debugging_features=false
 v8_enable_disassembler=true
 v8_enable_object_print=true
 v8_enable_pointer_compression=false
+v8_enable_webassembly=false
 v8_generate_external_defines_header=true
 v8_optimized_debug=true
 v8_postmortem_support=true
 v8_imminent_deprecation_warnings=false
 v8_deprecation_warnings=false
-use_cxx17=true
 cppgc_enable_young_generation=true
 '@
 }
@@ -69,6 +71,18 @@ if ($VerbosePreference -eq "SilentlyContinue") {
 }
 try {
     "V8 Build Starting" | timestamp | Write-Verbose
+    if ($Args.Count -gt 1) {
+        throw 'Only supports 1 argument. Usage: "BuildV8FromSource.ps1 [update]"'
+    }
+    if (($Args.Count -eq 1) -and ($Args[0] -cne "update")) {
+        throw 'Argument must be update. Usage: "BuildV8FromSource.ps1 [update]"'
+    }
+    "  Check if VS2022 is installed" | timestamp | Write-Verbose
+    $VSWhere = "C:\Program Files (x86)\Microsoft Visual Studio\Installer\vswhere.exe"
+    $InstallationPath = & $VSWhere -version '[17.0,18.0)' -latest -nologo -property installationPath
+    if ($LASTEXITCODE -ne 0) {
+        Throw "VS2022 may not be installed or vswhere.exe failed"
+    }
     "  Checking environmet variables" | timestamp | Write-Verbose
     if (!(Test-Path Env:\DEPOT_TOOLS_WIN_TOOLCHAIN)) {
         Throw "Missing environment variable DEPOT_TOOLS_WIN_TOOLCHAIN"
@@ -104,18 +118,22 @@ try {
     if ($PythonVersion.Split('.')[0] -ne "Python 3") {
         throw "Python 2 is needed (don't use Python 3 or Python 1)"
     }
-    "Removing old build directory" | timestamp | Write-Verbose
-    Remove-Item -ErrorAction Ignore -Recurse -Force C:\build
-    if (Test-Path C:\build) {
-        throw "Failed to delete C:\build directory"
+    if ($Args.Count -eq 0) {
+        "Removing old build directory" | timestamp | Write-Verbose
+        Remove-Item -ErrorAction Ignore -Recurse -Force C:\build
+        if (Test-Path C:\build) {
+            throw "Failed to delete C:\build directory"
+        }
+        New-Item -ItemType Directory C:\build | Out-Null
     }
-    New-Item -ItemType Directory C:\build | Out-Null
     Push-Location C:\build
     try {
-        "Fetching V8" | timestamp | Write-Verbose
-        cmd.exe /C "fetch v8 2>&1"
-        if ($LASTEXITCODE -ne 0) {
-            Throw "fetch v8 failed"
+        if ($Args.Count -eq 0) {
+            "Fetching V8" | timestamp | Write-Verbose
+            cmd.exe /C "fetch v8 2>&1"
+            if ($LASTEXITCODE -ne 0) {
+                Throw "fetch v8 failed"
+            }
         }
         Push-Location .\v8
         try {
@@ -133,10 +151,14 @@ try {
                 Throw "gcclient sync failed"
             }
             #
-            # Add define of CPPGC_YOUNG_GENERATION in cppgc_base_config
+            # BUILD.gn
             #
-            "Add define of CPPGC_YOUNG_GENERATION in cppgc_base_config" | timestamp | Write-Verbose
+            "Adjust BUILD.gn" | timestamp | Write-Verbose
             $FoundCppgcBaseConfig=$false
+            $FoundActionGenV8Gn=$false
+            $FoundSourceSetCppgcBase=$false
+            $FoundConfigInternalConfigBase=$false
+            $FoundConfigInternalConfig=$false
             (Get-Content BUILD.gn) |
                 Foreach-Object -process {
                     if ($_ -match '^config\("cppgc_base_config"\) {') {
@@ -151,11 +173,109 @@ try {
                     } elseif ($_ -match '^}') {
                         $FoundCppgcBaseConfig=$false
                         $_
+                    } elseif ($_ -match '^  action\("gen_v8_gn"\) {') {
+                        $FoundActionGenV8Gn=$true
+                        $_
+                    } elseif (($_ -match '^    visibility = \[ ":\*" \]') -and ($FoundActionGenV8Gn)) {
+                        $FoundActionGenV8Gn=$false
+                        '    visibility = ['
+                        '      ":*",'
+                        '      "tools\v8windbg\:*"'
+                        '    ]'
+                    } elseif ($_ -match '^v8_source_set\("cppgc_base"\) {') {
+                        $FoundSourceSetCppgcBase=$true
+                        $_
+                    } elseif (($_ -match '^}') -and ($FoundSourceSetCppgcBase)) {
+                        $FoundSourceSetCppgcBase=$false
+                        ''
+                        '  if (v8_generate_external_defines_header) {'
+                        '    sources += [ "$target_gen_dir/include/v8-gn.h" ]'
+                        '    include_dirs = [ "$target_gen_dir/include" ]'
+                        '    public_deps += [ ":gen_v8_gn" ]'
+                        '  }'
+                        $_
+                    } elseif ($_ -match '^config\("internal_config_base"\) {') {
+                        $FoundConfigInternalConfigBase=$true
+                        $_
+                    } elseif (($_ -match '^    "\$target_gen_dir",') -and ($FoundConfigInternalConfigBase)) {
+                        $FoundConfigInternalConfigBase=$false
+                        $_
+                        '    "$target_gen_dir/include",'
+                    } elseif ($_ -match '^config\("internal_config"\) {') {
+                        $FoundConfigInternalConfig=$true
+                        $_
+                    } elseif (($_ -match '^  defines = \[\]') -and ($FoundConfigInternalConfig)) {
+                        $FoundConfigInternalConfig=$false
+                        '  defines = ["_SILENCE_CXX20_OLD_SHARED_PTR_ATOMIC_SUPPORT_DEPRECATION_WARNING"]'
                     } else {
                         $_
                     }
                 } |
                 Set-Content BUILD.gn -Force
+            #
+            # tools\v8windbg\BUILD.gn
+            #
+            "Adjust v8 source set of target v8windbg_test in tools\v8windbg\BUILD.gn" | timestamp | Write-Verbose
+            $FoundV8SoutceSetV8windbgTest=$false
+            (Get-Content tools\v8windbg\BUILD.gn) |
+                Foreach-Object -process {
+                    if ($_ -match '^v8_source_set\("v8windbg_test"\) {') {
+                        $FoundV8SoutceSetV8windbgTest=$true
+                        $_
+                    } elseif (($_ -match '^}') -and ($FoundV8SoutceSetV8windbgTest)) {
+                        $FoundV8SoutceSetV8windbgTest=$false
+                        ''
+                        '  sources += [ "../../out/Debug/gen/include/v8-gn.h" ]'
+                        '  deps += [ "../..:gen_v8_gn" ]'
+                        $_
+                    } elseif ($_ -match '^config\("v8windbg_config"\) {') {
+                        $_
+                        '  configs = [ "../..:internal_config_base" ]'
+                    } else {
+                        $_
+                    }
+                } |
+                Set-Content tools\v8windbg\BUILD.gn -Force
+            #
+            # tools\gen-v8-gn.py
+            #
+            "Adjust Python script tools\gen-v8-gn.py" | timestamp | Write-Verbose
+            $SkipLines = 0
+            (Get-Content tools\gen-v8-gn.py) |
+                Foreach-Object -process {
+                    if ($_ -match '^def generate_positive_definition\(out, define\):') {
+                        $_
+                        '  if define.find("=") >= 0:'
+                        '    [define, value] = define.split("=")'
+                        '    out.write('''''''
+                        '#ifndef {define}'
+                        '#define {define} {value}'
+                        '#else'
+                        '#if {define} != {value}'
+                        '#error "{define} defined but not set to {value}"'
+                        '#endif'
+                        '#endif  // {define}'
+                        '''''''.format(define=define, value=value))'
+                        '  else:'
+                        '    out.write('''''''
+                        '#ifndef {define}'
+                        '#define {define} 1'
+                        '#else'
+                        '#if {define} != 1'
+                        '#error "{define} defined but not set to 1"'
+                        '#endif'
+                        '#endif  // {define}'
+                        '''''''.format(define=define))'
+                        $SkipLines = 9
+                    } else {
+                        if ($SkipLines -le 0) {
+                            $_
+                        } else {
+                            $SkipLines--
+                        }
+                    }
+                } |
+                Set-Content tools\gen-v8-gn.py -Force
             #
             # Debug
             #
@@ -201,10 +321,25 @@ try {
             New-Item -ErrorAction Ignore -ItemType Directory C:\build\v8\out\VisualStudio | Out-Null
             $ReleaseLines = GetV8ReleaseArgs
             Set-Content -Encoding Ascii -Path C:\build\v8\out\VisualStudio\args.gn -Value $ReleaseLines
-            cmd.exe /C "gn gen --ide=vs out\VisualStudio 2>&1"
+            cmd.exe /C "gn gen --ide=vs2022 out\VisualStudio 2>&1"
             if ($LASTEXITCODE -ne 0) {
                 Throw "gn gen failed for VisualStudio"
             }
+            #
+            # Fix age-table-unittest.cc
+            #
+            "Fix age-table-unittest.cc" | timestamp | Write-Verbose
+            (Get-Content test/unittests/heap/cppgc/age-table-unittest.cc) |
+                Foreach-Object -process {
+                    if ($_ -match '^  void\* heap_end = heap_start \+ kCagedHeapReservationSize - 1;') {
+                        '  void* heap_end = heap_start + api_constants::kCagedHeapDefaultReservationSize - 1;'
+                    } elseif ($_ -match '^      api_constants::kCagedHeapReservationSize \* 4\);') {
+                        '      api_constants::kCagedHeapDefaultReservationSize * 4);'
+                    } else {
+                        $_
+                    }
+                } |
+                Set-Content test/unittests/heap/cppgc/age-table-unittest.cc -Force
             #
             # Fix asm_to_inline_asm.py
             #
@@ -224,46 +359,6 @@ try {
                     }
                 } |
                 Set-Content $PathToasm_to_inline_asm
-            #
-            # Fix .\test\unittests\heap\shared-heap-unittest.cc
-            #
-            "Fix shared-heap-unittest.cc" | timestamp | Write-Verbose
-            Push-Location .\test\unittests\heap
-            try {
-                (Get-Content shared-heap-unittest.cc) |
-                    Foreach-Object -process {
-                        if ($_ -match '^  using ThreadType = TestType::ThreadType;')
-                        {
-                            '  using ThreadType = typename TestType::ThreadType;'
-                        } else {
-                            $_
-                        }
-                    } |
-                    Set-Content shared-heap-unittest.cc -Force
-            }
-            finally {
-                Pop-Location
-            }
-            #
-            # Fix .\include\v8-platform.h
-            #
-            "Fix v8-platform.h" | timestamp | Write-Verbose
-            Push-Location .\include
-            try {
-                (Get-Content v8-platform.h) |
-                    Foreach-Object -process {
-                        if ($_ -match '^    return floor\(CurrentClockTimeMillis\(\)\);')
-                        {
-                            '    return static_cast<int64_t>(floor(CurrentClockTimeMillis()));'
-                        } else {
-                            $_
-                        }
-                    } |
-                    Set-Content v8-platform.h -Force
-            }
-            finally {
-                Pop-Location
-            }
             #
             # Build Debug and Release
             #
